Add tag option 'skip if'
This option allow faker to skip a struct field if another field
is not empty.

This is useful if you have a struct that is a model for a database
table. Usually, for associations, you have a field with the foreign
key (for example AuthorId) and a field for the `Author` struct.

```
type Post {
  Id       uint64
  Title    string
  AuthorId uint64  `faker:"-"`
  Author   *Author `fake:"skip if AuthorId"`
}
```
To generate a new post with a new author just use `faker.Build(&post)`,
but sometimes you want to build a post without the `Author`
nested struct because you have set the `AuthorId` field.
In this case you can use the tag option `skip if`.